Reach 100% coverage by removing the last type-only file
c8 reported 99.35% line coverage because source/progress/event-payloads.ts
held only type aliases and the coverage gate is set to 100. Move every
payload type into progress-broadcaster.ts, the module that consumes them at
runtime, delete event-payloads.ts, and update the four consumers
(inspectors.ts, report-aggregator.ts, decorators.ts, elimination-emitter.ts).
The previously orphaned EliminationBundleResult export drops back to a
file-local type to satisfy knip.
